+++ src/dpr/components/async-report-list/utils.test.ts
@@ -5,9 +5,8 @@ import ColumnUtils from '../columns/utils'
 import PaginationUtils from '../pagination/utils'
 import Dict = NodeJS.Dict
 
-import createMockData from '../../../../test-app/mockAsyncData/mockAsyncData'
-
-import definitions from '../../../../test-app/mockAsyncData/mockReportDefinition'
+import createMockData from '../../../../test-app/mockClients/reports/mockAsyncData'
+import definitions from '../../../../test-app/mockClients/reports/mockReportDefinition'
 import { mockReportListRenderData } from '../../../../test-app/mockAsyncData/mockReportListRenderData'
 import { Columns } from '../columns/types'
 import { components } from '../../types/api'
