+++ src/dpr/utils/RequestReportUtils.test.ts
@@ -5,8 +5,8 @@ import RequestReportUtils from './RequestReportUtils'
 import { Services } from '../types/Services'
 import { ReportType } from '../types/UserReports'
 import { components } from '../types/api'
-import variant1 from '../../../test-app/mockAsyncData/mockVariants/variant1'
-import dashboardDefinitions from '../../../test-app/mockChartData/mockDashboardDefinition'
+import variant1 from '../../../test-app/mockClients/reports/mockVariants/variant1'
+import dashboardDefinitions from '../../../test-app/mockClients/dashboards/mockDashboardDefinition'
 import RequestedReportService from '../services/requestedReportService'
 import RecentlyViewedStoreService from '../services/recentlyViewedService'
