--- test-app/mockAsyncData/mockAsyncAppliedFilters.js
@@ -1,7 +1,0 @@
-const mockAppliedFilters = [
-  { name: 'filter1', value: 'dsfsdfsdf' },
-  { name: 'filter2', value: 'sdfsdf' },
-  { name: 'filter3', value: 'sdfsdf' },
-]
-
-module.exports = mockAppliedFilters
