+++ test-app/mockClients/reports/mockVariants/variant1.js
