+++ test-app/mockClients/reports/mockVariants/variant2.js
