+++ test-app/mockClients/reports/mockVariants/variant3.js
