+++ test-app/mockClients/store/mockRedisReportData.js
